pki: upgrade: do not overwrite apache certificate and key
Fix missing existence condition before duplicate the apache certificate
and key.

This will overwrite 3rd party certificate (if installed) at upgrade from
3.2 to more recent 3.2.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=969755

+++ packaging/fedora/setup/engine-upgrade.py
@@ -524,13 +524,14 @@ class CA():
         ):
             logging.debug("PKI: dup cert for apache")
             try:
-                utils.copyFile(
-                    src,
-                    dst,
-                    utils.getUsernameId("root"),
-                    utils.getGroupId("root"),
-                    0640
-                )
+                if not os.path.exists(dst):
+                    utils.copyFile(
+                        src,
+                        dst,
+                        utils.getUsernameId("root"),
+                        utils.getGroupId("root"),
+                        0640
+                    )
             except OSError:
                 logging.error("PKI: Cannot dup cert for apache")
                 raise
